Implement search bar functionality in main activity layout and connect to presenter

+++ AndroidProject/app/src/main/java/es/unican/movies/common/Utils.java
@@ -12,7 +12,6 @@ import java.lang.reflect.Type;
 import java.util.List;
 
 import es.unican.movies.model.Genre;
-import es.unican.movies.model.Movie;
 import es.unican.movies.model.Series;
 
 /**
@@ -48,7 +47,7 @@ public class Utils {
             return "-";
         }
 
-        double puntuacionNormalizada = 2*Math.log10(1 + voteCount);
+        double puntuacionNormalizada = 2*Math.log10(1 + (double) voteCount);
         double puntuacionSumaria = (voteAverage + puntuacionNormalizada) / 2;
         puntuacionSumaria = Math.round(puntuacionSumaria * 100.0) / 100.0; // Redondear a dos decimales
